requirements.txt
requests
Django
djangorestframework
PyJWT
djangorestframework-simplejwt
django-filter
djoser
